requirements.txt
selenium>=4.15.0
Pillow>=10.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0

